refactor: [jb-006] Review Dto Refactoring
Request에 대한 Dto를 분리하고, Dto에 있는 메타데이터를 제외,  json deserialize 오류 방지를 위한 NoArgsConstructor 어노테이션 추가
