+++ src/main/java/com/jjbacsa/jjbacsabackend/review/dto/ReviewDto.java
@@ -4,18 +4,18 @@ import com.jjbacsa.jjbacsabackend.shop.dto.ShopDto;
 import com.jjbacsa.jjbacsabackend.user.dto.UserDto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
-import java.time.LocalDateTime;
 
 
 @Data
 @AllArgsConstructor
+@NoArgsConstructor
 public class ReviewDto {
     private Long id;
     private UserDto userDto;
     private ShopDto shopDto;
     private String content;
     private int isTemp;
-    private LocalDateTime createdAt;
 
 }
