--- src/main/java/com/jjbacsa/jjbacsabackend/review/dto/ReviewRequest.java
@@ -1,14 +1,0 @@
-package com.jjbacsa.jjbacsabackend.review.dto;
-
-import lombok.*;
-
-@Getter
-@Setter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class ReviewRequest {
-
-    private String content;
-    private boolean isTemp;
-}
