+++ src/main/java/com/jjbacsa/jjbacsabackend/review/dto/request/ReviewRequest.java
@@ -1,22 +1,17 @@
-package com.jjbacsa.jjbacsabackend.review.dto;
+package com.jjbacsa.jjbacsabackend.review.dto.request;
 
 import com.jjbacsa.jjbacsabackend.review_image.dto.ReviewImageDto;
-import com.jjbacsa.jjbacsabackend.shop.dto.ShopDto;
-import com.jjbacsa.jjbacsabackend.user.dto.UserDto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.time.LocalDateTime;
 import java.util.List;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class ReviewWithImageDto {
-    private Long id;
-    private UserDto userDto;
-    private ShopDto shopDto;
+public class ReviewRequest {
+
     private String content;
     private int isTemp;
     private List<ReviewImageDto> reviewImageDtos;
