+++ src/main/java/com/jjbacsa/jjbacsabackend/review/dto/response/ReviewResponse.java
@@ -4,11 +4,13 @@ import com.jjbacsa.jjbacsabackend.shop.dto.ShopDto;
 import com.jjbacsa.jjbacsabackend.user.dto.UserDto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 import java.time.LocalDateTime;
 
 @Data
 @AllArgsConstructor
+@NoArgsConstructor
 public class ReviewResponse {
     private Long id;
     private UserDto userDto;
