+++ src/main/java/com/jjbacsa/jjbacsabackend/review/dto/response/ReviewWithImageResponse.java
@@ -5,12 +5,14 @@ import com.jjbacsa.jjbacsabackend.shop.dto.ShopDto;
 import com.jjbacsa.jjbacsabackend.user.dto.UserDto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 import java.time.LocalDateTime;
 import java.util.List;
 
 @Data
 @AllArgsConstructor
+@NoArgsConstructor
 public class ReviewWithImageResponse {
     private Long id;
     private UserDto userDto;
