+++ src/main/java/com/jjbacsa/jjbacsabackend/shop/dto/ShopDto.java
@@ -2,10 +2,12 @@ package com.jjbacsa.jjbacsabackend.shop.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 
 @Data
 @AllArgsConstructor
+@NoArgsConstructor
 public class ShopDto {
     private Long id;
     private String placeId;
