+++ src/main/java/com/jjbacsa/jjbacsabackend/user/dto/UserDto.java
@@ -2,9 +2,11 @@ package com.jjbacsa.jjbacsabackend.user.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Data
 @AllArgsConstructor
+@NoArgsConstructor
 public class UserDto {
     private Long id;
     private String account;
